Add basic controller functionality
Renames the old controller to GameRoom, and created a component named Controller that only holds the up/down logic.

The GameRoom connects to the socket and passes the channel down to the Controller.

Makes use of setInterval to continuously send events to the server. When the mouse press is released the interval is canceled.