Fix "Mirror Conf" not appearing on the first request
This happened due to the way Canvas work. Basically, the canvas would try to render text with the VT323 font, but the font was loaded async, so on the first request (without the font cached) the text would not be visible, or would have a default font. The fix? Well, just use webfontloader and wait for the google font to load and then render the react app.