Fix paddle positioning.
There are 2 issues with the paddle positioning. It uses the middle of
the rectangle's y axis to position it. I was accounting for the top.

For the x axis, 0 is the beginning of the rectangle (e.g for the circle,
it is the middle of it). I miscalculated and removed on paddleWidth / 2
when positioning both paddles. But the correct calculations should be to
subtract paddleWidth on the left paddle and do nothing on the right
paddle.