Use konva animations to improve performance
This PR makes use of the raw Konva API to create an animation that updates the game board frame by frame. Basically, I created an object with that synchronizes its state with the server (via a channel), then we start a Konva animation, that constantly reads this state and redraws the layer of the game board. To achieve that I extract the raw Konva objects with refs and pass them to the new GameAnimator class.

The Board component functionality it's just to draw the initial state of the game, extract the refs from the canvas and start the animation. Notice the shouldComponentUpdate = () => false, it makes the Board component "static", not triggering DOM updates when we receive info on the sockets.

Performance is noticeably better. This does not make use of raf though.