Add sponsor logos to background and fix board background
This PR removes the old canvas based background and instead adds a position absolute div with the background text.
Using this new background we add the sponsor logos. Every 5 seconds the sponsor logo rotates until we reach the end (starting the cycle anew).

The board resizer was also fixed. With the new animate method, the changes to the board width/height were being ignored. To fix this we just have to change the `data.on` handler to use the updated height/width props.