Add dialyxir type checking
This includes a `bin/nightly_build` script to check if is being run by
the scheduler.

Since dialyxir takes a long time to run, we shouldn't run it on every
build. Instead, this allows us to check if we are running on a semaphore
scheduled build and in those cases, run dialyxir.